restarting Lobby work to separate game events
