+++ server/socket/User.js
@@ -1,0 +1,13 @@
+class User {
+  constructor(data, socketId, users) {
+    this.id = data.id
+    this.email = data.email
+    this.username = data.username || data.email.split('@')[0]
+    this.socketId = socketId
+    this.gameId = ''
+
+    users[this.socketId] = this
+  }
+}
+
+module.exports = User
